Move prepare oracle to prepare.
