+++ qualtran/bloqs/block_encoding/lcu_block_encoding_test.py
@@ -27,7 +27,7 @@ from qualtran.bloqs.block_encoding.lcu_block_encoding import (
     BlackBoxPrepare,
     BlackBoxSelect,
 )
-from qualtran.bloqs.block_encoding.lcu_select_and_prepare import PrepareOracle, SelectOracle
+from qualtran.bloqs.state_preparation.prepare_base import PrepareOracle, SelectOracle
 
 
 def test_lcu_block_encoding(bloq_autotester):
