+++ qualtran/bloqs/block_encoding/lcu_select_and_prepare.py
@@ -67,48 +67,3 @@ _SELECT_ORACLE_DOC = BloqDocSpec(
     import_line='from qualtran.bloqs.block_encoding.lcu_select_and_prepare import SelectOracle',
     examples=[],
 )
-
-
-class PrepareOracle(GateWithRegisters):
-    r"""Abstract base class that defines the API for a PREPARE Oracle.
-
-    Given a set of coefficients $\{c_0, c_1, ..., c_{N - 1}\}$, the PREPARE oracle is used to encode
-    the coefficients as amplitudes of a state $|\Psi\rangle = \sum_{l=0}^{N-1} \sqrt{\frac{c_l}{\lambda}} |l\rangle$
-    where $\lambda = \sum_l |c_l|$, using a selection register $|l\rangle$. In order to prepare such
-    a state, the PREPARE circuit is also allowed to use a junk register that is entangled with
-    selection register.
-
-    Thus, the action of a PREPARE circuit on an input state $|0\rangle$ can be defined as:
-
-    $$
-        \mathrm{PREPARE} |0\rangle = \sum_{l=0}^{N-1} \sqrt{ \frac{c_l}{\lambda} } |l\rangle |\mathrm{junk}_l\rangle
-    $$
-    """
-
-    @property
-    @abc.abstractmethod
-    def selection_registers(self) -> Tuple[Register, ...]:
-        ...
-
-    @cached_property
-    def junk_registers(self) -> Tuple[Register, ...]:
-        return ()
-
-    @cached_property
-    def signature(self) -> Signature:
-        return Signature([*self.selection_registers, *self.junk_registers])
-
-    @property
-    def l1_norm_of_coeffs(self) -> Optional['SymbolicFloat']:
-        r"""Sum of the absolute values of coefficients $c_i$.
-
-        For LCU Hamiltonians, this is usually referred to as $\lambda$ in texts.
-        """
-        return None
-
-
-_PREPARE_ORACLE_DOC = BloqDocSpec(
-    bloq_cls=PrepareOracle,
-    import_line='from qualtran.bloqs.block_encoding.lcu_select_and_prepare import PrepareOracle',
-    examples=[],
-)
