+++ qualtran/bloqs/reflections/prepare_identity.py
@@ -20,7 +20,7 @@ from attrs import field, frozen
 from qualtran import bloq_example, BloqDocSpec, QAny, Register, Soquet
 from qualtran.bloqs.basic_gates import Identity
 from qualtran.bloqs.basic_gates.on_each import OnEach
-from qualtran.bloqs.block_encoding.lcu_block_encoding import PrepareOracle
+from qualtran.bloqs.state_preparation.prepare_base import PrepareOracle
 from qualtran.resource_counting.generalizers import ignore_split_join
 from qualtran.symbolics.types import SymbolicInt
