+++ qualtran/bloqs/reflections/reflection_using_prepare.py
@@ -33,7 +33,7 @@ from qualtran.resource_counting.generalizers import ignore_split_join
 from qualtran.symbolics.types import SymbolicInt
 
 if TYPE_CHECKING:
-    from qualtran.bloqs.block_encoding.lcu_select_and_prepare import PrepareOracle
+    from qualtran.bloqs.state_preparation.prepare_base import PrepareOracle
     from qualtran.resource_counting import BloqCountT, SympySymbolAllocator
